+++ romancal/photom/__init__.py
@@ -1,0 +1,3 @@
+from .photom_step import PhotomStep
+
+__all__ = ['PhotomStep']