+++ CHANGES.rst
@@ -12,6 +12,12 @@ Documentation
 - Add documentation for DNS build 0.5, e.g. reference array trimming [#457]
 
 
+photom
+------
+
+ - Added photom correction step and unit tests. [#469]
+
+
 0.6.0 (2022-03-02)
 ==================